after Itai's comments July 8
